+++ BPt/extensions/Loaders.py
@@ -2,12 +2,15 @@ from sklearn.base import BaseEstimator, TransformerMixin
 import numpy as np
 import warnings
 import networkx as nx
+from scipy.linalg import lstsq
 
 
 class Identity(BaseEstimator, TransformerMixin):
 
     def __init__(self):
-        '''This loader simply flatten the input array and passed it along'''
+        '''This loader will simply flatten the input object, if not already.
+        This loader is used to for example pass along loaded surfaces
+        or volumes to a PCA or other simmilar transformation.'''
         pass
 
     def fit(self, X, y=None):
@@ -29,12 +32,15 @@ def load_surf(surf):
     # If str, assume file path
     if isinstance(surf, str):
 
-        try:
-            surf = np.load(surf)
-            return surf
-        except ValueError:
-            pass
+        # Only try to load with numpy if str ends with numpy ext
+        if surf.endswith('.npy') or surf.endswith('.npz'):
+            try:
+                surf = np.load(surf)
+                return surf
+            except ValueError:
+                pass
 
+        # If not numpy, or numpy failed, try nilearn load surf data
         try:
             from nilearn.surface import load_surf_data
         except ImportError:
@@ -49,13 +55,13 @@ def load_surf(surf):
     elif surf is None:
         return None
 
-    # Otherwise assume either valid array-like passed or a Parc object
-    else:
+    # If a Parc object
+    elif hasattr(surf, 'get_parc'):
+        return surf.get_parc(copy=True)
 
-        try:
-            return surf.get_parc(copy=True)
-        except AttributeError:
-            return np.array(surf).copy()
+    # Otherwise assume either valid array-like passed
+    else:
+        return np.array(surf).copy()
 
 
 class SurfLabels(BaseEstimator, TransformerMixin):
@@ -82,26 +88,43 @@ class SurfLabels(BaseEstimator, TransformerMixin):
             nilearn.surface.load_surf_data(), which you
             will need nilearn installed to use.)
 
-        background_labels : int, array-like of int or None, optional
-            This parameter determines which label, if any, in the corresponding
-            passed labels, should be treated as 'background' and therefore no ROI
-            calculated for that value or values. You may pass either a single interger
-            value, an array-like of integer values, or None, to calculate ROIs for everything.
+        background_labels : int, array-like of int, optional
+            This parameter determines which label, if any,
+            in the corresponding
+            passed labels, should be treated as 'background'
+            and therefore no ROI
+            calculated for that value or values.
+            You may pass either a single interger
+            value, an array-like of integer values.
 
-            (default = 0)
+            If not background label is desired, just pass
+            a label which doesn't exist in any of the data,
+            e.g., -100.
+
+            ::
+
+                default = 0
 
         mask : None, str or array-like, optional
             This parameter allows you to optional pass a mask of values in
-            which to not calculate ROI values for. This can be passed as a str or
-            array-like of values (just like labels), and should be comprised of
-            a boolean array (or 1's and 0's), where a value of 1 means that value
-            should be kept, and a value of 0, for that value should be masked away.
+            which to not calculate ROI values for.
+            This can be passed as a str or
+            array-like of values (just like labels),
+            and should be comprised of
+            a boolean array (or 1's and 0's),
+            where a value of 1 means that value
+            will be ignored (set to background label)
+            should be kept, and a value of 0,
+            for that value should be masked away.
             This array should have the same length as the passed `labels`.
 
-            (default = None)
+            ::
+
+                default = None
 
         strategy: specific str, custom_func, optional
-            This parameter dictates the function to be applied to each data's ROI's
+            This parameter dictates the function to be applied
+            to each data's ROI's
             individually, e.g., mean to calculate the mean by ROI.
 
             If a str is passed, it must correspond to one of the below preset
@@ -137,14 +160,18 @@ class SurfLabels(BaseEstimator, TransformerMixin):
             [data_dim, n_contrasts], data_dim = 0, and lastly for a 1D
             array, data_dim is also 0.
 
-            (default = 'mean')
+            ::
+
+                default = 'mean'
 
         vectorize : bool, optional
             If the returned array should be flattened to 1D. E.g., if the
             last step in a set of loader steps this should be True, if before
             a different step it may make sense to set to False.
 
-            (default = True)
+            ::
+
+                default = True
         '''
 
         self.labels = labels
@@ -163,8 +190,13 @@ class SurfLabels(BaseEstimator, TransformerMixin):
 
         # Mask labels if mask
         if self.mask_ is not None:
-            self.mask_ = self.mask_.astype(bool)
-            self.labels_[self.mask_] = self.background_label
+
+            # Raise error if wrong shapes
+            if len(self.mask_) != len(self.labels_):
+                raise RuntimeError('length of mask must have '
+                                   'the same length / shape as '
+                                   'the length of labels!')
+            self.labels_[self.mask_.astype(bool)] = self.background_label
 
         # X can either be a 1D surface, or a 2D surface
         # (e.g. - for timeseries or stacked contrasts)
@@ -179,10 +211,6 @@ class SurfLabels(BaseEstimator, TransformerMixin):
         # Proc self.background_label, if int turn to np array
         if isinstance(self.background_label, int):
             self.background_label_ = np.array([self.background_label])
-
-        # If None...
-        elif self.background_label is None:
-            self.background_label_ = np.array([])
 
         # Otherwise, if already list-like, just cast to np array
         else:
@@ -223,10 +251,12 @@ class SurfLabels(BaseEstimator, TransformerMixin):
     def transform(self, X):
         ''' If X has the both the same dimensions, raise warning'''
 
+        self._check_fitted()
+
         if len(X.shape) == 2 and (X.shape[0] == X.shape[1]):
             warnings.warn('X was passed with the same length',
                           ' in each dimension, ',
-                          'Assuming that axis=0 is that data dimension',
+                          'Assuming that axis=0 is the data dimension',
                           ' w/ vertex values')
 
         # The data dimension is just the dimension with
@@ -254,28 +284,272 @@ class SurfLabels(BaseEstimator, TransformerMixin):
         if not self.vectorize:
             return X_trans
 
-        self.o_shape_ = X_trans.shape
+        # Save original shape if vectorize called,
+        # used for reverse transform
+        self.original_shape_ = X_trans.shape
         return X_trans.flatten()
 
-    def inverse_transform(self, X):
+    def inverse_transform(self, X_trans):
 
         # Reverse the vectorize
         if self.vectorize:
-            X = X.reshape(self.o_shape_)
+            X_trans = X_trans.reshape(self.original_shape_)
 
-        X_trans = np.zeros(self.X_shape_, dtype=X.dtype, order='C')
+        X = np.zeros(self.X_shape_, dtype=X_trans.dtype, order='C')
 
         if self.data_dim_ == 1:
             X_trans = np.rollaxis(X_trans, -1)
             X = np.rollaxis(X, -1)
 
         for i, label in enumerate(self.non_bkg_unique_):
-            X_trans[self.labels_ == label] = X[i]
+            X[self.labels_ == label] = X_trans[i]
 
         if self.data_dim_ == 1:
-            X_trans = np.rollaxis(X_trans, -1)
+            X = np.rollaxis(X, -1)
 
-        return X_trans
+        return X
+
+
+class SurfMaps(BaseEstimator, TransformerMixin):
+
+    def __init__(self, maps, strategy='auto', mask=None, vectorize=True):
+        '''Simmilar to NiftiMapsMasker from nilearn
+        but for surfaces, and designed to work with BPt Loader.
+
+        This object calculates the signal for each of the
+        passed maps as extracted from the input during fit,
+        and returns for each map a value.
+
+        maps : str or array-like, optional
+            This parameter represents the maps in which
+            to apply to each surface, where the shape of
+            the passed maps should be (# of vertex, # of maps)
+            or in other words, the size of the data array in the first
+            dimension and the number of maps
+            (i.e., the number of outputted ROIs from fit)
+            as the second dimension.
+
+            You may pass maps as either an array-like,
+            or the str file location of a numpy or other
+            valid surface file format array in which to load.
+
+        strategy : {'auto', 'ls', 'average'}, optional
+            The stratgey in which the maps are used to extract
+            signal. If 'ls' is selected, which stands for
+            least squares, the least-squares solution will
+            be used for each region.
+
+            Alternatively, if 'average' is passed, then
+            the weighted average value for each map
+            will be computed.
+
+            By default 'auto' will be selected,
+            which will use 'average' if the passed
+            maps contain only positive weights, and
+            'ls' in the case that there are
+            any negative values in the passed maps.
+
+            Otherwise, you can set a specific strategy.
+            In deciding which method to use,
+            consider an example. Let's say the
+            fit data X, and maps are
+
+            ::
+
+                data = np.array([1, 1, 5, 5])
+                maps = np.array([[0, 0],
+                                 [0, 0],
+                                 [1, -1],
+                                 [1, -1]])
+
+            In this case, the 'ls' method would
+            yield region signals [2.5, -2.5], whereas
+            the weighted 'average' method, would yield
+            [5, 5], notably ignoring the negative weights.
+            This highlights an important limitation to the
+            weighted averaged method, as it does not
+            handle negative values well.
+
+            On the other hand, consider changing the maps
+            weights to
+
+            ::
+
+                data = np.array([1, 1, 5, 5])
+                maps = np.array([[0, 1],
+                                 [0, 2],
+                                 [1, 0],
+                                 [1, 0]])
+
+                ls_sol = [5. , 0.6]
+                average_sol = [5, 1]
+
+            In this case, we can see that the weighted
+            average gives a maybe more intuative summary
+            of the regions. In general, it depends on
+            what signal you are trying to summarize, and
+            how you are trying to summarize it.
+
+        mask : None, str or array-like, optional
+            This parameter allows you to optional pass a mask of values in
+            which to not calculate ROI values for.
+            This can be passed as a str or
+            array-like of values (just like maps),
+            and should be comprised of
+            a boolean array (or 1's and 0's),
+            where a value of 1 means that value
+            will be ignored (set to 0)
+            should be kept, and a value of 0,
+            for that value should be masked away.
+            This array should have the same length as the passed `maps`.
+            Specifically, where the shape of maps is (size, n_maps),
+            the shape of mask should be (size).
+
+            ::
+
+                default = None
+
+        vectorize : bool, optional
+            If the returned array should be flattened to 1D. E.g., if this is
+            the last step in a set of loader steps this should be True.
+            Also note, if the surface data it is being applied to is 1D,
+            then the output will be 1D regardless of this parameter.
+
+            ::
+
+                default = True
+        '''
+
+        self.maps = maps
+        self.strategy = strategy
+        self.mask = mask
+        self.vectorize = vectorize
+
+    def fit(self, X, y=None):
+
+        # Load mask if any
+        self.mask_ = load_surf(self.mask)
+
+        # Load maps
+        self.maps_ = load_surf(self.maps)
+
+        # Make the maps if passed mask set to 0 in those spots
+        if self.mask_ is not None:
+
+            # Raise error if wrong shapes
+            if len(self.mask_) != self.maps_.shape[0]:
+                raise RuntimeError('length of mask must have '
+                                   'the same length / shape as '
+                                   'the first dimension of passed '
+                                   'maps!')
+            self.maps_[self.mask_.astype(bool)] = 0
+
+        # X can either be a 1D surface, or a 2D surface
+        # (e.g. - for timeseries or stacked contrasts)
+        if len(X.shape) > 2:
+            raise RuntimeError('X can be at most 2D.')
+
+        # Check to make sure dimension of data is correct
+        if self.maps_.shape[0] not in X.shape:
+            raise RuntimeError('Size of labels not found in X. '
+                               'Make sure your data is in the same '
+                               'space as the labels you are using!')
+
+        # Make sure strategy exists
+        if self.strategy not in ['auto', 'ls', 'average']:
+            raise RuntimeError('strategy must be '
+                               '"ls", "average" or "auto"!')
+
+        return self
+
+    def fit_transform(self, X, y=None):
+        return self.fit(X, y).transform(X)
+
+    def _check_fitted(self):
+        if not hasattr(self, "maps_"):
+            raise ValueError('It seems that SurfMaps has not been fitted. '
+                             'You must call fit() before calling transform()')
+
+    def transform(self, X):
+
+        self._check_fitted()
+
+        if len(X.shape) == 2 and (X.shape[0] == X.shape[1]):
+            warnings.warn('X was passed with the same length',
+                          ' in each dimension, ',
+                          'Assuming that axis=0 is the data dimension',
+                          ' w/ vertex values')
+
+        # The data dimension is just the dimension with
+        # the first dimension of maps
+        self.data_dim_ = X.shape.index(self.maps_.shape[0])
+        self.X_shape_ = X.shape
+
+        # If data in second dimension, transpose
+        if self.data_dim_ == 1:
+            X = X.T
+
+        # Set strategy if auto
+        strategy = self.strategy
+        if strategy == 'auto':
+            strategy = 'ls'
+            if np.all(self.maps_ >= 0):
+                strategy = 'average'
+
+        # Run the correct transform based on strategy
+        if strategy == 'ls':
+            X_trans = self._transform_ls(X)
+        elif strategy == 'average':
+            X_trans = self._transform_average(X)
+        else:
+            X_trans = None
+
+        # Always return as shape of extra data if any by
+        # number of maps, with number of maps as last dimension
+
+        # Return based on vectorize
+        if not self.vectorize:
+            return X_trans
+
+        # Save original transformed output shape if vectorize
+        self.original_shape_ = X_trans.shape
+        return X_trans.flatten()
+
+    def _transform_ls(self, X):
+        '''X should be data points, or data points x stacked.'''
+
+        X_trans = lstsq(self.maps_, X)[0]
+        return X_trans.T
+
+    def _transform_average(self, X):
+        '''X should be data points, or data points x stacked.'''
+
+        X_trans = []
+
+        # For each map - take weighted average
+        for m in range(self.maps_.shape[1]):
+
+            try:
+                X_trans.append(np.average(X, axis=0, weights=self.maps_[:, m]))
+            except ZeroDivisionError:
+                pass
+
+        return np.array(X_trans).T
+
+    def inverse_transform(self, X_trans):
+
+        # Reverse the vectorize, if needed
+        if self.vectorize:
+            X_trans = X_trans.reshape(self.original_shape_)
+
+        # For ls should be something like
+        # np.dot(region_signals, maps.T)
+
+        # For inverse transform weighted average,
+        # make sure to include check for sum(weights) == 0
+        # and to fill in that spot with zeros
+
+        raise RuntimeError('Not Implemented Yet')
 
 
 # Create wrapper for nilearn connectivity measure to make it
@@ -366,14 +640,19 @@ class Networks(BaseEstimator, TransformerMixin):
                      'global_eff': nx.global_efficiency,
                      'local_eff': nx.local_efficiency,
                      'transitivity': nx.transitivity,
-                     'avg_eigenvector_centrality': self._avg_eigenvector_centrality,
-                     'avg_closeness_centrality': self._avg_closeness_centrality,
+                     'avg_eigenvector_centrality':
+                     self._avg_eigenvector_centrality,
+                     'avg_closeness_centrality':
+                     self._avg_closeness_centrality,
                      'avg_degree': self._avg_degree,
                      'avg_triangles': self._avg_triangles,
                      'avg_pagerank': self._avg_pagerank,
-                     'avg_betweenness_centrality': self._avg_betweenness_centrality,
-                     'avg_information_centrality': self._avg_information_centrality,
-                     'avg_shortest_path_length': nx.average_shortest_path_length
+                     'avg_betweenness_centrality':
+                     self._avg_betweenness_centrality,
+                     'avg_information_centrality':
+                     self._avg_information_centrality,
+                     'avg_shortest_path_length':
+                     nx.average_shortest_path_length
                      }
 
         X_trans = []
@@ -382,7 +661,6 @@ class Networks(BaseEstimator, TransformerMixin):
         return np.array(X_trans)
 
     # Local
-
     def _avg_degree(self, G):
         avg_degree = np.mean([i[1] for i in nx.degree(G)])
         return avg_degree
@@ -392,8 +670,9 @@ class Networks(BaseEstimator, TransformerMixin):
         return avg_triangles
 
     def _avg_eigenvector_centrality(self, G):
-        avg_eigenvector_centrality = np.mean([nx.eigenvector_centrality_numpy(G)[
-                                             i] for i in nx.eigenvector_centrality_numpy(G)])
+        avg_eigenvector_centrality =\
+            np.mean([nx.eigenvector_centrality_numpy(G)[
+                     i] for i in nx.eigenvector_centrality_numpy(G)])
         return avg_eigenvector_centrality
 
     def _avg_closeness_centrality(self, G):
@@ -403,12 +682,14 @@ class Networks(BaseEstimator, TransformerMixin):
 
     def _avg_betweenness_centrality(self, G):
         avg_betweenness_centrality = np.mean(
-            [nx.betweenness_centrality(G)[i] for i in nx.betweenness_centrality(G)])
+            [nx.betweenness_centrality(G)[i]
+             for i in nx.betweenness_centrality(G)])
         return avg_betweenness_centrality
 
     def _avg_information_centrality(self, G):
         avg_information_centrality = np.mean(
-            [nx.information_centrality(G)[i] for i in nx.information_centrality(G)])
+            [nx.information_centrality(G)[i]
+             for i in nx.information_centrality(G)])
         return avg_information_centrality
 
     def _avg_pagerank(self, G):