+++ BPt/pipeline/Imputers.py
@@ -18,7 +18,6 @@ def get_imputer_and_params(imputer_str, extra_params, params, search_type,
                            random_state=None, num_feat_keys=None):
 
     imputer, extra_imputer_params, imputer_params =\
-        get_obj_and_params(imputer_str, IMPUTERS, extra_params, params,
-                           search_type)
+        get_obj_and_params(imputer_str, IMPUTERS, extra_params, params)
 
     return imputer(**extra_imputer_params), imputer_params