+++ BPt/pipeline/Models.py
@@ -207,14 +207,27 @@ try:
 except ImportError:
     pass
 
+try:
+    from ..extensions.AutoGluon import AutoGluon
+
+    AVALIABLE['binary']['auto gluon'] = 'auto gluon binary'
+    AVALIABLE['categorical']['auto gluon'] = 'auto gluon categorical'
+    AVALIABLE['regression']['auto gluon'] = 'auto gluon regression'
+
+    MODELS['auto gluon binary'] = (AutoGluon, ['pt binary'])
+    MODELS['auto gluon categorical'] = (AutoGluon, ['pt multiclass'])
+    MODELS['auto gluon regression'] = (AutoGluon, ['pt regression'])
+
+except ImportError:
+    pass
+
 
 def get_base_model_and_params(model_type, extra_params, model_type_params,
                               search_type, random_state=None,
                               num_feat_keys=None):
 
     model, extra_model_params, model_type_params =\
-        get_obj_and_params(model_type, MODELS, extra_params,
-                           model_type_params, search_type)
+        get_obj_and_params(model_type, MODELS, extra_params, model_type_params)
 
     # Init model, w/ any user passed params + class params
     model = model(**extra_model_params)