+++ BPt/pipeline/Pipeline_Pieces.py
@@ -166,12 +166,10 @@ class Pieces():
             process_params_by_type(obj=user_obj,
                                    obj_str=name,
                                    base_params=deepcopy(param),
-                                   extra_params=extra_params,
-                                   search_type=self.spec['search_type'])
+                                   extra_params=extra_params)
 
         # If passing a user object, kind of stupid to pass default, non search
-        # params
-        # via a dict..., but hey give it a try
+        # params, via a dict..., but hey give it a try
         try:
             user_obj.set_params(**extra_user_obj_params)
         except AttributeError:
@@ -197,15 +195,11 @@ class Pieces():
 
     def _check_params(self, obj):
 
-        try:
-            obj.random_state = self.spec['random_state']
-        except AttributeError:
-            pass
+        if hasattr(obj, 'random_state'):
+            setattr(obj, 'random_state', self.spec['random_state'])
 
-        try:
-            obj.n_jobs = self.spec['n_jobs']
-        except AttributeError:
-            pass
+        if hasattr(obj, 'n_jobs'):
+            setattr(obj, 'n_jobs', self.spec['n_jobs'])
 
         return obj